Fix minimal build.
